+++ src/zarr.driver.c
@@ -39,6 +39,12 @@ struct Storage*
 compressed_zarr_v2_zstd_init();
 struct Storage*
 compressed_zarr_v2_lz4_init();
+struct Storage*
+zarr_v3_init();
+struct Storage*
+compressed_zarr_v3_zstd_init();
+struct Storage*
+compressed_zarr_v3_lz4_init();
 
 //
 //                  GLOBALS
@@ -49,6 +55,9 @@ enum StorageKind
     Storage_Zarr,
     Storage_ZarrBlosc1ZstdByteShuffle,
     Storage_ZarrBlosc1Lz4ByteShuffle,
+    Storage_ZarrV3,
+    Storage_ZarrV3Blosc1ZstdByteShuffle,
+    Storage_ZarrV3Blosc1Lz4ByteShuffle,
     Storage_Number_Of_Kinds
 };
 
@@ -71,6 +80,9 @@ storage_kind_to_string(const enum StorageKind kind)
         CASE(Storage_Zarr);
         CASE(Storage_ZarrBlosc1ZstdByteShuffle);
         CASE(Storage_ZarrBlosc1Lz4ByteShuffle);
+        CASE(Storage_ZarrV3);
+        CASE(Storage_ZarrV3Blosc1ZstdByteShuffle);
+        CASE(Storage_ZarrV3Blosc1Lz4ByteShuffle);
 #undef CASE
         default:
             return "(unknown)";
@@ -99,6 +111,9 @@ zarr_describe(const struct Driver* driver,
         XXX(Zarr),
         XXX(ZarrBlosc1ZstdByteShuffle),
         XXX(ZarrBlosc1Lz4ByteShuffle),
+        XXX(ZarrV3),
+        XXX(ZarrV3Blosc1ZstdByteShuffle),
+        XXX(ZarrV3Blosc1Lz4ByteShuffle),
     };
     // clang-format on
 #undef XXX
@@ -160,6 +175,10 @@ acquire_driver_init_v0(acquire_reporter_t reporter)
             [Storage_Zarr] = zarr_v2_init,
             [Storage_ZarrBlosc1ZstdByteShuffle] = compressed_zarr_v2_zstd_init,
             [Storage_ZarrBlosc1Lz4ByteShuffle] = compressed_zarr_v2_lz4_init,
+            [Storage_ZarrV3] = zarr_v3_init,
+            [Storage_ZarrV3Blosc1ZstdByteShuffle] =
+              compressed_zarr_v3_zstd_init,
+            [Storage_ZarrV3Blosc1Lz4ByteShuffle] = compressed_zarr_v3_lz4_init,
         };
         memcpy(
           globals.constructors, impls, nbytes); // cppcheck-suppress uninitvar